+++ elements/linuxmodule/perfcountaccum.hh
@@ -54,8 +54,8 @@ class PerfCountAccum : public PerfCountUser { public:
  private:
   
   int _which;
-  unsigned long long _accum;
-  unsigned long long _npackets;
+  uint64_t _accum;
+  uint64_t _npackets;
 
   static String read_handler(Element *, void *);
   static int reset_handler(const String &, Element *, void *, ErrorHandler *);
